[EuiDataGridBody] Split up into custom and virtualized body renderers
Split up parent body into a basic ternary, and move tests accordingly

Custom renderer logic will be in a separate commit (to hopefully make DRYing things out easier to follow)
