+++ src/components/datagrid/body/data_grid_body.test.tsx
@@ -7,137 +7,62 @@
  */
 
 import React from 'react';
-import { mount, render, shallow } from 'enzyme';
+import { render } from '../../../test/rtl';
 
 import { RowHeightUtils } from '../utils/__mocks__/row_heights';
 import { schemaDetectors } from '../utils/data_grid_schema';
 
-import { EuiDataGridBody, Cell } from './data_grid_body';
+import { EuiDataGridBody } from './data_grid_body';
+
+// Body props, reused by other body unit tests
+export const dataGridBodyProps = {
+  headerIsInteractive: true,
+  rowCount: 1,
+  visibleRows: { startRow: 0, endRow: 1, visibleRowCount: 1 },
+  columnWidths: { columnA: 100 },
+  columns: [
+    { id: 'columnA', schema: 'boolean' },
+    { id: 'columnB', isExpandable: true },
+  ],
+  leadingControlColumns: [],
+  trailingControlColumns: [],
+  visibleColCount: 2,
+  schema: {
+    columnA: { columnType: 'boolean' },
+    columnB: { columnType: 'string' },
+  },
+  renderCellValue: () => <span>cell content</span>,
+  interactiveCellId: 'someId',
+  inMemory: { level: 'enhancements' as any },
+  inMemoryValues: {},
+  handleHeaderMutation: jest.fn(),
+  setVisibleColumns: jest.fn(),
+  switchColumnPos: jest.fn(),
+  schemaDetectors,
+  rowHeightUtils: new RowHeightUtils(),
+  isFullScreen: false,
+  gridStyles: {},
+  gridWidth: 300,
+  gridRef: { current: null },
+  gridItemsRendered: { current: null },
+  wrapperRef: { current: document.createElement('div') },
+};
 
 describe('EuiDataGridBody', () => {
-  const requiredProps = {
-    headerIsInteractive: true,
-    rowCount: 1,
-    visibleRows: { startRow: 0, endRow: 1, visibleRowCount: 1 },
-    columnWidths: { columnA: 20 },
-    columns: [
-      { id: 'columnA', schema: 'boolean' },
-      { id: 'columnB', isExpandable: true },
-    ],
-    leadingControlColumns: [],
-    trailingControlColumns: [],
-    visibleColCount: 2,
-    schema: {
-      columnA: { columnType: 'boolean' },
-      columnB: { columnType: 'string' },
-    },
-    renderCellValue: () => <span>cell content</span>,
-    interactiveCellId: 'someId',
-    inMemory: { level: 'enhancements' as any },
-    inMemoryValues: {},
-    handleHeaderMutation: jest.fn(),
-    setVisibleColumns: jest.fn(),
-    switchColumnPos: jest.fn(),
-    schemaDetectors,
-    rowHeightUtils: new RowHeightUtils(),
-    isFullScreen: false,
-    gridStyles: {},
-    gridWidth: 300,
-    gridRef: { current: null },
-    gridItemsRendered: { current: null },
-    wrapperRef: { current: document.createElement('div') },
-  };
-
-  beforeAll(() => {
-    Object.defineProperty(HTMLElement.prototype, 'offsetHeight', {
-      configurable: true,
-      value: 34,
-    });
+  it('renders a virtualized data grid body by default', () => {
+    const { container } = render(<EuiDataGridBody {...dataGridBodyProps} />);
+    expect(container.querySelector('.euiDataGrid__virtualized')).toBeTruthy();
   });
 
-  it('renders', () => {
-    // EuiDataGridBody should be `render`ed here over `mount` due to large
-    // snapshot memory issues
-    const component = render(<EuiDataGridBody {...requiredProps} />);
-    expect(component).toMatchSnapshot();
-    expect(component.find('[data-test-subj="dataGridRowCell"]')).toHaveLength(
-      2
-    );
-  });
-
-  it('renders leading columns, trailing columns, and footer rows', () => {
-    const component = mount(
+  it('renders a custom rendered data grid body if `renderCustomGridBody` is passed', () => {
+    const { container } = render(
       <EuiDataGridBody
-        {...requiredProps}
-        leadingControlColumns={[
-          {
-            id: 'someLeadingColumn',
-            headerCellRender: () => <div />,
-            rowCellRender: () => <div />,
-            width: 30,
-          },
-        ]}
-        trailingControlColumns={[
-          {
-            id: 'someTrailingColumn',
-            headerCellRender: () => <div />,
-            rowCellRender: () => <div />,
-            width: 40,
-          },
-        ]}
-        visibleColCount={4}
-        renderFooterCellValue={() => <footer data-test-subj="footer" />}
+        {...dataGridBodyProps}
+        renderCustomGridBody={() => null}
       />
     );
-    expect(component.find('Cell')).toHaveLength(4);
-    expect(component.find('[data-test-subj="footer"]')).toHaveLength(2);
-  });
-
-  it('passes some virtualization options to the underlying react-window grid', () => {
-    const onItemsRendered = jest.fn();
-    const component = mount(
-      <EuiDataGridBody
-        {...requiredProps}
-        virtualizationOptions={{
-          initialScrollTop: 50,
-          className: 'test',
-          onItemsRendered,
-        }}
-      />
-    );
-    expect(component.find('.test').exists()).toBe(true);
-    expect(onItemsRendered).toHaveBeenCalled();
-  });
-
-  // TODO: Test final height/weights in Cypress
-
-  // TODO: Test tabbing in Cypress
-
-  // TODO: Test column resizing in Cypress
-});
-
-describe('Cell', () => {
-  const requiredProps = {
-    columnIndex: 0,
-    rowIndex: 0,
-    style: {},
-    data: {
-      rowMap: {},
-      rowOffset: 0,
-      leadingControlColumns: [],
-      trailingControlColumns: [],
-      columns: [{ id: 'C' }],
-      columnWidths: {},
-      defaultColumnWidth: 30,
-      schema: {},
-      schemaDetectors,
-      interactiveCellId: '',
-      renderCellValue: jest.fn(),
-    },
-  };
-
-  it('is a light wrapper around EuiDataGridCell', () => {
-    const component = shallow(<Cell {...requiredProps} />);
-    expect(component.find('EuiDataGridCell').exists()).toBe(true);
+    expect(
+      container.querySelector('.euiDataGrid__customRenderBody')
+    ).toBeTruthy();
   });
 });
