+++ src/components/datagrid/body/data_grid_body_custom.tsx
@@ -1,0 +1,15 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import React, { FunctionComponent } from 'react';
+
+import { EuiDataGridBodyProps } from '../data_grid_types';
+
+export const EuiDataGridBodyCustomRender: FunctionComponent<EuiDataGridBodyProps> = () => {
+  return <div className="euiDataGrid__customRenderBody">TODO</div>;
+};
